Widen cross sections

Widen or narrow a collection of cross sections. xt_widen_by() widens (narrows) by unit length; xt_widen_times() widens (narrows) by a multiplicative factor.

Usage

xt_widen_by(object, by)

xt_widen_times(object, times)

Arguments

object

A cross section object.

by

Amount to widen the channel by, using units in common with the cross sectional units. Negative values will narrow cross sections. Either a vector of length equal to the number of cross sections, or length 1.

times

Multiplicative factor to widen the channel by. Values less than 1 will narrow cross sections. Either a vector of length equal to the number of cross sections, or length 1.